test: extract public api tests to type and runtime tests
